Refactored the markup code
