+++ www/main/settings/utils/index.js
@@ -1,3 +1,1 @@
-export { default as generateSwitchMarkup } from './generate-switch-markup';
-export { default as generateDropdownMarkup } from './generate-dropdown-markup';
-export { default as generateRangeMarkup } from './generate-range-markup';
+export { default as generateMarkup } from './generate-markup';
