+++ www/src/scss/_settings.scss
@@ -45,9 +45,9 @@
         align-items: center;
         border-radius: 4px;
         display: flex;
-        min-height: 40px;
         justify-content: space-between;
         margin: 4px 0;
+        min-height: 40px;
         padding: 6px;
         transition: 0.2s background-color ease;
